backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
apscheduler
bson